# perf(url): avoid trivial reallocs