Reducing url parsing benches by ~15%:

Before:
```
test long  ... bench:         484 ns/iter (+/- 6) = 88 MB/s
test short ... bench:         347 ns/iter (+/- 5) = 72 MB/s
```

After:
```
test long  ... bench:         413 ns/iter (+/- 4) = 104 MB/s
test short ... bench:         298 ns/iter (+/- 9) = 83 MB/s
```